requirements.txt
# API
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-dotenv==1.0.0

# WhisperX e dependências existentes...
whisperx==3.3.0
ctranslate2==4.4.0
faster-whisper==1.1.0
nltk
transformers

# PyAnnote.Audio e dependências
pyannote.audio==3.3.2
asteroid-filterbanks>=0.4
einops>=0.6.0
huggingface-hub>=0.13.0
lightning>=2.0.1
omegaconf>=2.1,<3.0
pyannote.core>=5.0.0
pyannote.database>=5.0.1
pyannote.metrics>=3.2
pyannote.pipeline>=3.0.1
pytorch-metric-learning>=2.1.0
rich>=12.0.0
semver>=3.0.0
speechbrain>=1.0.0
tensorboardX>=2.6
torch-audiomentations>=0.11.0
torchmetrics>=0.11.0

# PyTorch
torch>=2.0.0 
torchaudio>=2.0.0
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html

# Utilitários
pandas>=2.0.0
numpy>=1.24.0
soundfile>=0.13.0
librosa>=0.10.0
colorama